+++ beam/animations/__init__.py
@@ -6,3 +6,5 @@ from .light import Light
 from .rainbow import Rainbow
 from .strip import Strip
 from .zap import Zap
+from .color_wipe_rotate import ColorWipeRotate
+from .color_wipe_sequential import ColorWipeSequential
