+++ beam/animations/color_wipe_rotate.py
@@ -1,5 +1,7 @@
 import random
 from animations.base import BaseBeamAnim, check_interrupt, adjustable
+import bibliopixel.colors as color_util
+from beam.state import beam_state
 
 
 class ColorWipeRotate(BaseBeamAnim):
