+++ beam/animations/langtons_ant.py
@@ -1,5 +1,8 @@
 import random
 from animations.base import BaseBeamAnim, check_interrupt, adjustable
+from beam.state import beam_state
+import beam.util as beam_util
+import bibliopixel.colors as color_util
 
 
 class LangtonsAnt(BaseBeamAnim):
