+++ beam/animations/light.py
@@ -1,4 +1,6 @@
+from beam.state import beam_state
 from animations.base import BaseBeamAnim, check_interrupt, adjustable
+import beam.util as beam_util
 
 
 class Light(BaseBeamAnim):
@@ -11,7 +13,7 @@ class Light(BaseBeamAnim):
     @adjustable
     def step(self, amt=1):
         for x, y in self.grid():
-            c = beam_state.colors[get_location(x, y) % len(beam_state.colors)]
+            c = beam_state.colors[beam_util.get_location(x, y) % len(beam_state.colors)]
             self.layout.set(x, y, c)
 
         self._step = 0
