+++ example.env
@@ -1,5 +1,5 @@
 {
-    "PIXELS_PER_STRIP": "40",
-    "NUM_STRIPS": "15",
+    "PIXELS_PER_STRIP": 40,
+    "NUM_STRIPS": 15,
     "ENV": "dev"
 }
